requirements.txt
astroid==2.4.1
autopep8==1.5.2
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
google-api-core==1.16.0
google-auth==1.10.2
google-cloud-automl==0.9.0
google-cloud-bigquery==1.23.1
google-cloud-core==1.2.0
google-cloud-documentai==0.1.0
google-cloud-language==1.3.0
google-cloud-storage==1.25.0
google-cloud-vision==0.41.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
idna==2.8
img2pdf==0.3.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.0.0
proto-plus==0.4.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pylint==2.5.2
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.12.1
